third_party/blink/web_tests/fast/forms/enctype-attribute-expected.txt

Test enctype and formenctype attributes

Missing value default:
PASS form1.enctype is "application/x-www-form-urlencoded"
PASS form1.getAttribute("enctype") is null


Invalid value default:
PASS form1.enctype = "foobar"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.getAttribute("enctype") is "foobar"
PASS form1.setAttribute("enctype", "baz"); form1.enctype is "application/x-www-form-urlencoded"


Text and its subtypes:
PASS form1.enctype = "text"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "plain"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/1d-interleaved-parityfec"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/calendar"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/css"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/csv"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/dns"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/enriched"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/example"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/fwdred"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/html"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/rfc822-headers"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/richtext"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/rtx"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/sgml"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/t140"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/troff"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/ulpfec"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/uri-list"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/vcard"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/xml"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "text/xml-external-parsed-entity"; form1.enctype is "application/x-www-form-urlencoded"


Multipart and its subtypes:
PASS form1.enctype = "multipart"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "form-data"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/mixed"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/message"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/digest"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/alternative"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/alternative"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/byteranges"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/encrypted"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/example"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/parallel"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/related"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/report"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.enctype = "multipart/signed"; form1.enctype is "application/x-www-form-urlencoded"


Valid values:
PASS form1.enctype = "Text/Plain"; form1.enctype is "text/plain"
PASS form1.getAttribute("enctype") is "Text/Plain"
PASS form1.setAttribute("enctype", "Text/Plain"); form1.enctype is "text/plain"
PASS form1.enctype = "Multipart/Form-Data"; form1.enctype is "multipart/form-data"
PASS form1.getAttribute("enctype") is "Multipart/Form-Data"
PASS form1.setAttribute("enctype", "Multipart/Form-Data"); form1.enctype is "multipart/form-data"


Values with whitespace:
PASS form1.enctype = "  multipart/form-data"; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.getAttribute("enctype") is "  multipart/form-data"
PASS form1.enctype = "multipart/form-data  "; form1.enctype is "application/x-www-form-urlencoded"
PASS form1.getAttribute("enctype") is "multipart/form-data  "
PASS form1.enctype = "application/x-www-form-urlencoded "; form1.enctype is "application/x-www-form-urlencoded"


PASS successfullyParsed is true

TEST COMPLETE
